Added automatic deactivation of mods
Mods which fail to be polled a number of times, i.e. the version
information about the mod could not be retrieved several times in a row,
are now deactivated and put into a special list automatically. This list
can be looked up with the 'failedmods' command. Using the 'poll' command
on a mod removes it from the special list. A mod that has been put on
the automatic list can thus be activated until it fails the check a few
times again.